+++ src/ui/views/DangerZone.tsx
@@ -1,6 +1,8 @@
 import React, { useState } from "react";
+import { PHASE } from "../../common";
+import type { View } from "../../common/types";
 import useTitleBar from "../hooks/useTitleBar";
-import { toWorker } from "../util";
+import { helpers, toWorker } from "../util";
 
 const AutoSave = ({ autoSave }: { autoSave: boolean }) => {
 	const [processing, setProcessing] = useState(false);
@@ -65,7 +67,7 @@ const AutoSave = ({ autoSave }: { autoSave: boolean }) => {
 	);
 };
 
-const DangerZone = ({ autoSave }: { autoSave: boolean }) => {
+const DangerZone = ({ autoSave, godMode, phase }: View<"dangerZone">) => {
 	useTitleBar({
 		title: "Danger Zone",
 	});
@@ -80,7 +82,7 @@ const DangerZone = ({ autoSave }: { autoSave: boolean }) => {
 					drastic measures might save it.
 				</p>
 
-				<div className="btn-group mb-3">
+				<div className="btn-group mb-5">
 					<button
 						type="button"
 						className="btn btn-light-bordered"
@@ -118,6 +120,87 @@ const DangerZone = ({ autoSave }: { autoSave: boolean }) => {
 						Preseason
 					</button>
 				</div>
+
+				<h2>Trade deadline</h2>
+
+				<p>
+					This will not sim any games, it will just toggle whether the trade
+					deadline has passed or not this season, and delete any scheduled trade
+					deadline later this season.
+				</p>
+
+				{!godMode ? (
+					<p className="text-warning">
+						This feature is only available in{" "}
+						<a href={helpers.leagueUrl(["god_mode"])}>God Mode</a>.
+					</p>
+				) : phase !== PHASE.REGULAR_SEASON &&
+				  phase !== PHASE.AFTER_TRADE_DEADLINE ? (
+					<p className="text-warning">
+						This only works during the regular season.
+					</p>
+				) : null}
+
+				<div className="mb-5">
+					{phase === PHASE.AFTER_TRADE_DEADLINE ? (
+						<button
+							type="button"
+							className="btn btn-god-mode border-0"
+							disabled={!godMode}
+							onClick={() => {
+								toWorker("main", "toggleTradeDeadline");
+							}}
+						>
+							Switch to before trade deadline
+						</button>
+					) : (
+						<button
+							type="button"
+							className="btn btn-god-mode border-0"
+							disabled={phase !== PHASE.REGULAR_SEASON || !godMode}
+							onClick={() => {
+								toWorker("main", "toggleTradeDeadline");
+							}}
+						>
+							Switch to after trade deadline
+						</button>
+					)}
+				</div>
+
+				<h2>All-Star Game</h2>
+
+				<p>
+					If the All-Star Game has not yet happened, you can move it up to right
+					now, so that it will happen before the next currently scheduled game.
+					This also works if the current season has no All-Star Game - it will
+					add one, and it will happen before the next game.
+				</p>
+
+				<p>
+					If the All-Star Game has already happened and you add another one... I
+					guess you get an extra All-Star Game?
+				</p>
+
+				{phase !== PHASE.REGULAR_SEASON &&
+				phase !== PHASE.AFTER_TRADE_DEADLINE ? (
+					<p className="text-warning">
+						This only works during the regular season.
+					</p>
+				) : null}
+
+				<button
+					type="button"
+					className="btn btn-light-bordered mb-5"
+					disabled={
+						phase !== PHASE.REGULAR_SEASON &&
+						phase !== PHASE.AFTER_TRADE_DEADLINE
+					}
+					onClick={() => {
+						toWorker("main", "allStarGameNow");
+					}}
+				>
+					Schedule All-Star Game now
+				</button>
 			</div>
 			<div className="col-md-6">
 				<h2>Auto save</h2>