+++ src/worker/views/dangerZone.ts
@@ -1,8 +1,10 @@
-import { local } from "../util";
+import { g, local } from "../util";
 
 const updateDangerZone = async () => {
 	return {
 		autoSave: local.autoSave,
+		godMode: g.get("godMode"),
+		phase: g.get("phase"),
 	};
 };